fix up tuntap for go 1.13 netpoller
running the tuntap in go 1.13 resulted Interface.Read() returning
an "not pollable" error (from the runtime's poll.ErrNotPollable).
This, it turns out, is due to /dev/net/tun in linux not being pollable
(in the epoll sense) until after the TUNSETIFF ioctl has been done.

The right fix, done here, is to open /dev/net/tun as a raw file
descriptor, and ioctl it before constructing an *os.File which gets
added to the poll set when Read() is called.

 See https://github.com/golang/go/issues/30426
 and https://github.com/golang/go/issues/30624

and go source code commit a5fdd58c84b6b0a1ae5a53faebc0550024e3a066
which adds ErrNotPollable and exposes this error which otherwise
was getting silently thrown away.

This code works properly on the AP, too (master branch, using go 1.12.9,
but it should work a long way back)